requirements.txt
pandas==1.3.3
numpy==1.21.2
ta==0.10.2
requests==2.26.0
python-telegram-bot==20.0
binance-connector==3.3.0
